modularize elements
