+++ packages/react-hig/src/adapters/prepareUpdate.js
