+++ packages/react-hig/src/elements/Button.js
@@ -1,0 +1,48 @@
+/**
+Copyright 2016 Autodesk,Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+
+*/
+import * as HIG from 'hig.web';
+
+import HIGElement from './HIGElement';
+import createComponent from '../adapters/createComponent';
+
+class Button extends HIGElement {
+  constructor(initialProps) {
+    super(HIG.Button, initialProps);
+  }
+
+  commitUpdate(updatePayload, oldProps, newProps) {
+    const mapping = {
+      title: 'setTitle',
+      link: 'setLink'
+    };
+
+    for (let i = 0; i < updatePayload.length; i += 2) {
+      const propKey = updatePayload[i];
+      const propValue = updatePayload[i + 1];
+
+      if (mapping[propKey]) {
+        this.hig[mapping[propKey]](propValue);
+      } else {
+        this.commitPropChange(propKey, propValue);
+      }
+    }
+  }
+}
+
+const ButtonComponent = createComponent(Button);
+
+export default ButtonComponent;
