+++ packages/react-hig/src/elements/Button.test.js
@@ -1,8 +1,24 @@
+/**
+Copyright 2016 Autodesk,Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+
+*/
 import { mount } from 'enzyme';
 import * as HIG from 'hig.web';
 import React from 'react';
 
-import { Button } from './react-hig';
+import Button from './Button';
 
 describe('<Button>', () => {
   /**
