+++ packages/react-hig/src/elements/GlobalNav/GlobalNav.test.js
@@ -1,8 +1,24 @@
+/**
+Copyright 2016 Autodesk,Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+
+*/
 import { mount } from 'enzyme';
 import * as HIG from 'hig.web';
 import React from 'react';
 
-import { GlobalNav } from './react-hig';
+import GlobalNav from './GlobalNav';
 
 describe('<GlobalNav>', () => {
   /**
@@ -121,7 +137,7 @@ describe('<GlobalNav>', () => {
         const wrapper = mount(
           <GlobalNav>
             <GlobalNav.SideNav>
-              <GlobalNav.SideNav.Sections />
+              <GlobalNav.SideNav.SectionList />
             </GlobalNav.SideNav>
           </GlobalNav>,
           {
@@ -141,8 +157,8 @@ describe('<GlobalNav>', () => {
           mount(
             <GlobalNav>
               <GlobalNav.SideNav>
-                <GlobalNav.SideNav.Sections />
-                <GlobalNav.SideNav.Sections />
+                <GlobalNav.SideNav.SectionList />
+                <GlobalNav.SideNav.SectionList />
               </GlobalNav.SideNav>
             </GlobalNav>
           );
@@ -169,9 +185,9 @@ describe('<GlobalNav>', () => {
           const wrapper = mount(
             <GlobalNav>
               <GlobalNav.SideNav>
-                <GlobalNav.SideNav.Sections>
-                  <GlobalNav.SideNav.Sections.Item {...sectionDefaults} />
-                </GlobalNav.SideNav.Sections>
+                <GlobalNav.SideNav.SectionList>
+                  <GlobalNav.SideNav.SectionList.Item {...sectionDefaults} />
+                </GlobalNav.SideNav.SectionList>
 
               </GlobalNav.SideNav>
             </GlobalNav>,
@@ -215,10 +231,10 @@ describe('<GlobalNav>', () => {
           const wrapper = mount(
             <GlobalNav>
               <GlobalNav.SideNav>
-                <GlobalNav.SideNav.Sections>
-                  <GlobalNav.SideNav.Sections.Item {...section1Defaults} />
-                  <GlobalNav.SideNav.Sections.Item {...section2Defaults} />
-                </GlobalNav.SideNav.Sections>
+                <GlobalNav.SideNav.SectionList>
+                  <GlobalNav.SideNav.SectionList.Item {...section1Defaults} />
+                  <GlobalNav.SideNav.SectionList.Item {...section2Defaults} />
+                </GlobalNav.SideNav.SectionList>
               </GlobalNav.SideNav>
             </GlobalNav>,
             {
@@ -271,13 +287,15 @@ describe('<GlobalNav>', () => {
               return (
                 <GlobalNav>
                   <GlobalNav.SideNav>
-                    <GlobalNav.SideNav.Sections>
+                    <GlobalNav.SideNav.SectionList>
                       {this.state.showingItem &&
-                        <GlobalNav.SideNav.Sections.Item
+                        <GlobalNav.SideNav.SectionList.Item
                           {...section1Defaults}
                         />}
-                      <GlobalNav.SideNav.Sections.Item {...section2Defaults} />
-                    </GlobalNav.SideNav.Sections>
+                      <GlobalNav.SideNav.SectionList.Item
+                        {...section2Defaults}
+                      />
+                    </GlobalNav.SideNav.SectionList>
                   </GlobalNav.SideNav>
 
                 </GlobalNav>
