+++ packages/react-hig/src/elements/GlobalNav/Item.js
@@ -1,0 +1,73 @@
+/**
+Copyright 2016 Autodesk,Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+
+*/
+import createComponent from '../../adapters/createComponent';
+import HIGElement from '../HIGElement';
+
+export class Item extends HIGElement {
+  commitUpdate(updatePayload, oldProps, newProps) {
+    for (let i = 0; i < updatePayload.length; i += 2) {
+      const propKey = updatePayload[i];
+      const propValue = updatePayload[i + 1];
+
+      switch (propKey) {
+        case 'icon':
+          this.hig.setIcon(propValue);
+          break;
+        case 'title':
+          this.hig.setTitle(propValue);
+          break;
+        case 'link':
+          this.hig.setLink(propValue);
+          break;
+        case 'onClick': {
+          const dispose = this._disposeFunctions.get('onClickDispose');
+
+          if (dispose) {
+            dispose();
+          }
+
+          this._disposeFunctions.set(
+            'onClickDispose',
+            this.hig.onClick(propValue)
+          );
+          break;
+        }
+
+        case 'onHover': {
+          const dispose = this._disposeFunctions.get('onHoverDispose');
+
+          if (dispose) {
+            dispose();
+          }
+
+          this._disposeFunctions.set(
+            'onHoverDispose',
+            this.hig.onHover(propValue)
+          );
+          break;
+        }
+        default: {
+          console.warn(`${propKey} is unknown`);
+        }
+      }
+    }
+  }
+}
+
+const ItemComponent = createComponent(Item);
+
+export default ItemComponent;
