+++ packages/react-hig/src/interface/partitionProps.js
